Update to 0.3.0 (26.1.2)

+++ src/main/kotlin/com/odtheking/odinaddon/features/impl/dungeon/DungeonMobs.kt
@@ -1,23 +1,13 @@
 package com.odtheking.odinaddon.features.impl.dungeon
 
-import com.odtheking.odin.clickgui.settings.Setting.Companion.withDependency
-import com.odtheking.odin.clickgui.settings.impl.BooleanSetting
 import com.odtheking.odin.clickgui.settings.impl.ColorSetting
-import com.odtheking.odin.events.core.on
 import com.odtheking.odin.features.Module
 import com.odtheking.odin.utils.Colors
-import com.odtheking.odin.utils.equalsOneOf
-import com.odtheking.odin.utils.handlers.TickTask
-import com.odtheking.odin.utils.matchesOneOf
-import com.odtheking.odin.utils.modMessage
 import com.odtheking.odin.utils.render.textDim
 import com.odtheking.odin.utils.skyblock.dungeon.DungeonUtils
-import com.odtheking.odinaddon.features.impl.skyblock.event.EntityWorldEvent
 import com.odtheking.odinaddon.utils.EntityCollection
 import com.odtheking.odinaddon.utils.getRoomData
-import net.minecraft.world.entity.LivingEntity
 import net.minecraft.world.entity.decoration.ArmorStand
-import net.minecraft.world.entity.monster.Giant
 import kotlin.collections.find
 
 object DungeonMobs : Module(
@@ -40,7 +30,7 @@ object DungeonMobs : Module(
         //TODO other blood mob names/types (golden, stealthy, etc.)
         DungeonUtils.inDungeons &&
                 it is ArmorStand &&
-                getRoomData(it.x, it.z)?.name == "Blood" &&
+                getRoomData(it.x.toInt(), it.z.toInt())?.name == "Blood" &&
                 (it.customName?.string?.contains("Giant") ?: false)
    } )
